Fresh polynomials

+++ content/numerical/NumberTheoreticTransform.h
@@ -45,7 +45,7 @@ vll conv(const vll &a, const vll &b) {
 	vll L(a), R(b), out(n);
 	L.resize(n), R.resize(n);
 	ntt(L), ntt(R);
-	FOR(i,0,n) out[-i & (n-1)] = (ll)L[i]*R[i] %mod * inv %mod;
+	FOR(i,0,n) out[-i & (n-1)] = L[i] * R[i] % mod * inv % mod;
 	ntt(out);
 	return {out.begin(), out.begin() + s};
 }
